api/requirements.txt
fastapi>=0.115.0
mangum>=0.17.0
setuptools>=65.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.9.0
supabase>=2.5.0
requests>=2.32.0
huggingface_hub>=0.29.0
moviepy>=1.0.3
Pillow>=10.4.0
imageio-ffmpeg>=0.4.9
gradio_client>=1.5.0
numpy>=1.24.0
